+++ arpeggio/__init__.py
@@ -169,6 +169,9 @@ class ParsingExpression(object):
         suppress (bool): If this is set to True than no ParseTreeNode will be
             created for this ParsingExpression. Default False.
     """
+
+    suppress = False
+
     def __init__(self, *elements, **kwargs):
 
         if len(elements) == 1:
@@ -183,7 +186,8 @@ class ParsingExpression(object):
             nodes = [nodes]
         self.nodes = nodes
 
-        self.suppress = kwargs.get('suppress', False)
+        if 'suppress' in kwargs:
+            self.suppress = kwargs['suppress']
 
         # Memoization. Every node cache the parsing results for the given input
         # positions.
@@ -714,8 +718,8 @@ class Match(ParsingExpression):
     """
     Base class for all classes that will try to match something from the input.
     """
-    def __init__(self, rule_name, root=False):
-        super(Match, self).__init__(rule_name=rule_name, root=root)
+    def __init__(self, rule_name, root=False, **kwargs):
+        super(Match, self).__init__(rule_name=rule_name, root=root, **kwargs)
 
     @property
     def name(self):
@@ -806,8 +810,9 @@ class RegExMatch(Match):
 
     '''
     def __init__(self, to_match, rule_name='', root=False, ignore_case=None,
-                 multiline=None, str_repr=None, re_flags=re.MULTILINE):
-        super(RegExMatch, self).__init__(rule_name, root)
+                 multiline=None, str_repr=None, re_flags=re.MULTILINE,
+                 **kwargs):
+        super(RegExMatch, self).__init__(rule_name, root, **kwargs)
         self.to_match_regex = to_match
         self.ignore_case = ignore_case
         self.multiline = multiline
@@ -860,8 +865,9 @@ class StrMatch(Match):
         ignore_case(bool): If case insensitive match is needed.
             Default is None to support propagation from global parser setting.
     """
-    def __init__(self, to_match, rule_name='', root=False, ignore_case=None):
-        super(StrMatch, self).__init__(rule_name, root)
+    def __init__(self, to_match, rule_name='', root=False, ignore_case=None,
+                 **kwargs):
+        super(StrMatch, self).__init__(rule_name, root, **kwargs)
         self.to_match = to_match
         self.ignore_case = ignore_case
 
@@ -903,6 +909,7 @@ class StrMatch(Match):
 
     def __hash__(self):
         return hash(self.to_match)
+
 
 
 # HACK: Kwd class is a bit hackish. Need to find a better way to
@@ -1020,6 +1027,10 @@ class ParseTreeNode(object):
             # If default actions are enabled
             return visitor.visit__default__(self, children)
 
+    def tree_str(self, indent=0):
+        return '{}{} [{}-{}]'.format('  ' * indent, self.rule.name,
+                                     self.position, self.position_end)
+
 
 class Terminal(ParseTreeNode):
     """
@@ -1068,6 +1079,10 @@ class Terminal(ParseTreeNode):
 
     def __repr__(self):
         return self.desc
+
+    def tree_str(self, indent=0):
+        return '{}: {}'.format(super(Terminal, self).tree_str(indent),
+                               self.value)
 
     def __eq__(self, other):
         return text(self) == text(other)
@@ -1127,6 +1142,11 @@ class NonTerminal(ParseTreeNode, list):
 
     def __repr__(self):
         return "[ %s ]" % ", ".join([repr(x) for x in self])
+
+    def tree_str(self, indent=0):
+        return '{}\n{}'.format(super(NonTerminal, self).tree_str(indent),
+                               '\n'.join([c.tree_str(indent + 1)
+                                          for c in self]))
 
     def __getattr__(self, rule_name):
         """
@@ -1718,7 +1738,8 @@ class CrossRef(object):
 
 class ParserPython(Parser):
 
-    def __init__(self, language_def, comment_def=None, *args, **kwargs):
+    def __init__(self, language_def, comment_def=None, syntax_classes=None,
+                 *args, **kwargs):
         """
         Constructs parser from python statements and expressions.
 
@@ -1727,8 +1748,13 @@ class ParserPython(Parser):
                 the root rule of the grammar.
             comment_def (python function): A python function that defines
                 the root rule of the comments grammar.
+            syntax_classes (dict): Overrides of special syntax parser
+                expression classes (StrMatch, Sequence, OrderedChoice).
+
         """
         super(ParserPython, self).__init__(*args, **kwargs)
+
+        self.syntax_classes = syntax_classes if syntax_classes else {}
 
         # PEG Abstract Syntax Graph
         self.parser_model = self._from_python(language_def)
@@ -1762,6 +1788,10 @@ class ParserPython(Parser):
         __rule_cache = {"EndOfFile": EndOfFile()}
         __for_resolving = []  # Expressions that needs crossref resolvnih
         self.__cross_refs = 0
+        _StrMatch = self.syntax_classes.get('StrMatch', StrMatch)
+        _OrderedChoice = self.syntax_classes.get('OrderedChoice',
+                                                 OrderedChoice)
+        _Sequence = self.syntax_classes.get('Sequence', Sequence)
 
         def inner_from_python(expression):
             retval = None
@@ -1802,9 +1832,10 @@ class ParserPython(Parser):
                     self.dprint("New rule: {} -> {}"
                                 .format(rule_name, retval.__class__.__name__))
 
-            elif type(expression) is text or isinstance(expression, StrMatch):
+            elif type(expression) is text or isinstance(expression, _StrMatch):
                 if type(expression) is text:
-                    retval = StrMatch(expression, ignore_case=self.ignore_case)
+                    retval = _StrMatch(expression,
+                                       ignore_case=self.ignore_case)
                 else:
                     retval = expression
                     if expression.ignore_case is None:
@@ -1839,7 +1870,7 @@ class ParserPython(Parser):
                 if any((isinstance(x, CrossRef) for x in retval.nodes)):
                     __for_resolving.append(retval)
 
-            elif isinstance(expression, Sequence) or \
+            elif isinstance(expression, _Sequence) or \
                     isinstance(expression, Repetition) or \
                     isinstance(expression, SyntaxPredicate) or \
                     isinstance(expression, Decorator):
@@ -1850,9 +1881,9 @@ class ParserPython(Parser):
 
             elif type(expression) in [list, tuple]:
                 if type(expression) is list:
-                    retval = OrderedChoice(expression)
+                    retval = _OrderedChoice(expression)
                 else:
-                    retval = Sequence(expression)
+                    retval = _Sequence(expression)
 
                 retval.nodes = [inner_from_python(e) for e in expression]
                 if any((isinstance(x, CrossRef) for x in retval.nodes)):